fix: store graphLookup depthField as int32 to match MongoDB BSON wire type (do-pavg)
Executed-By: dongo/polecats/topaz
Rig: dongo
Role: polecats

+++ internal/handler/common/aggregations/stages/graph_lookup.go
@@ -283,7 +283,7 @@ func (gl *graphLookup) traverse(doc *types.Document, fromDocs []*types.Document)
 
 				result := fromDoc.DeepCopy()
 				if gl.depthField != "" {
-					result.Set(gl.depthField, currentDepth)
+					result.Set(gl.depthField, int32(currentDepth))
 				}
 
 				results = append(results, result)
